web/requirements.txt
flask
pyjwt
flask_sqlalchemy
requests
six
authlib
python-dotenv